ref: extract YesNoOption into its own file
